+++ src/symbol.rs
@@ -62,7 +62,7 @@ impl<'a> Symbol<'a> {
             Symbol::Field(m, p) => Some(format!("{}::{}", p.name, m.name)),
             Symbol::EnumElement(el, e) => Some(format!("{}::{}", e.name, el.name)),
             Symbol::Type(ast::Type {
-                kind: ast::TypeKind::Resolved(qualified_name, _),
+                kind: ast::TypeKind::ResolvedItem(qualified_name, _),
                 ..
             }) => Some(qualified_name.clone()),
             Symbol::Type(_) => None,